% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_description.R
\name{.update_description}
\alias{.update_description}
\title{Update \code{DESCRIPTION} to include all packages required to compile Stan code.}
\usage{
.update_description(pkgdir, auto_config = FALSE, msg = TRUE)
}
\arguments{
\item{pkgdir}{Package root directory.}

\item{auto_config}{Whether or not package should configure itself.  This means importing \code{rstantools} for linking via \code{configure[.win]}) and setting \code{Biarch: true} in the \code{DESCRIPTION}.}

\item{msg}{Whether or not to display a message if an attempt to update \code{DESCRIPTION} is made.}
}
\value{
Whether or not \code{DESCRIPTION} file was modified.
}
\description{
Update \code{DESCRIPTION} to include all packages required to compile Stan code.
}
\details{
Sanitizes the package dependencies as follows:
\itemize{
  \item If \code{rstantools} dependency version is lower, don't update existing dependency.
  \item If \code{rstantools} \code{Imports} dependency already exists in \code{Depends}, keep the \code{Depends} package but increase version if needed.
}

Issues the following messages:
\itemize{
  \item If \code{auto_config = FALSE} and \code{rstantools} is imported.
  \item If \code{auto_config = FALSE} and \code{Biarch: true}.
}
}
\keyword{internal}
