% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cpp.R
\name{init_cpp}
\alias{init_cpp}
\title{Register functions implemented in C++}
\usage{
init_cpp(name, path)
}
\arguments{
\item{name}{The name of your package as a string.}

\item{path}{The path to the root directory for your package as a string. If
not specified it is assumed that this is already the current working
directory.}
}
\value{
This function is only called for its side effect of writing the
  necessary "init.cpp" file to the package's \code{src/} directory.
}
\description{
If you set up your package using \code{\link{rstan_package_skeleton}} before
version version 1.2.1 of \pkg{rstantools} it may be necessary for you to call
this function yourself. If you used \code{\link{rstan_package_skeleton}} in
version 1.2.1 or later then this has already been done automatically.
}
