% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Generic function for accessing the pointwise log-likelihood}
\usage{
log_lik(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{log_lik} methods should return a \eqn{S} by \eqn{N} matrix,
  where \eqn{S} is the size of the posterior sample and \eqn{N} is the number
  of data points.
}
\description{
See \code{\link[rstanarm]{log_lik.stanreg}} in the
\pkg{\link{rstanarm}} package for an example.
}

