% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsamples.R
\name{nsamples}
\alias{nsamples}
\title{Generic function for extracting the number of posterior samples}
\usage{
nsamples(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\description{
Extract the number of posterior samples stored in a fitted Bayesian model.
}
