% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\name{posterior_interval}
\alias{posterior_interval}
\title{Generic function for computing posterior uncertainty intervals}
\usage{
posterior_interval(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{posterior_interval} methods should return a matrix with two
  columns and as many rows as model parameters (or a subset of parameters
  specified by the user). For a given probability, \eqn{p}, the columns
  correspond to the lower and upper \eqn{100p}\% interval limits and have the
  names \eqn{100\alpha/2}\% and \eqn{100(1 - \alpha/2)}\%, where \eqn{\alpha
  = 1-p}. For example, for \eqn{90}\% intervals, the column names should be
  \code{"5\%"} and \code{"95\%"}.
}
\description{
These intervals are often referred to as credible intervals, but we use the
term uncertainty intervals to highlight the fact that wider intervals
correspond to greater uncertainty. See
\code{\link[rstanarm]{posterior_interval.stanreg}} in the
\pkg{\link{rstanarm}} package for an example.
}

