% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Generic function for drawing from the posterior predictive distribution}
\usage{
posterior_predict(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{posterior_predict} methods should return a \eqn{D} by \eqn{N}
  matrix, where \eqn{D} is the number of draws from the posterior predictive
  distribution and \eqn{N} is the sample size (number of observations) per
  draw.
}
\description{
See \code{\link[rstanarm]{posterior_predict.stanreg}} in the
\pkg{\link{rstanarm}} package for an example.
}

