% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sample.R
\name{posterior_sample}
\alias{posterior_sample}
\title{Generic function for accessing posterior draws}
\usage{
posterior_sample(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{posterior_sample} methods should return a a \eqn{S} by \eqn{P}
  matrix (with all chains combined), where \eqn{S} is the size of the
  posterior sample (or sample size specified by the user) and \eqn{P} is the
  number of parameters (or subset of parameters specified by the user).
}
\description{
\code{posterior_sample} is for quickly accessing the posterior draws of
parameters in matrix form with all chains combined. See
\code{\link[rstanarm]{posterior_sample.stanreg}} in the \pkg{\link{rstanarm}}
package for an example.
}

