% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_error.R
\name{predictive_error}
\alias{predictive_error}
\title{Generic function for computing in-sample or out-of-sample predictive errors}
\usage{
predictive_error(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{predictive_error} methods should return a \code{ndraws} by
  \code{nobs} matrix.
}
\description{
Generic function for computing predictive errors \eqn{y - y^{rep}}{y - yrep}
(in-sample, for observed \eqn{y}) or \eqn{y - \tilde{y}}{y - ytilde}
(out-of-sample, for new or held-out \eqn{y}). See
\code{\link[rstanarm]{predictive_error.stanreg}} in the \pkg{\link{rstanarm}}
package for an example.
}

