% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary}
\alias{prior_summary}
\alias{prior_summary.default}
\title{Generic function for extracting information about prior distributions}
\usage{
prior_summary(object, ...)

\method{prior_summary}{default}(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\value{
\code{prior_summary} methods should return an object containing
  information about the prior distribution(s) used for the given model.
  The structure of this object will depend on the method.

  The default method just returns \code{object$prior.info}, which is
  \code{NULL} if there is no \code{'prior.info'} element.
}
\description{
See \code{\link[rstanarm]{prior_summary.stanreg}} in the
\pkg{\link[rstanarm]{rstanarm}} package for an example.
}
\examples{
# See help("prior_summary", package = "rstanarm")

}
\seealso{
\itemize{
  \item The \pkg{\link[rstanarm]{rstanarm}} package for example methods
  (\href{https://CRAN.R-project.org/package=rstanarm}{CRAN},
  \href{https://github.com/stan-dev/rstanarm}{GitHub}).
}

\itemize{
  \item The guidelines for developers of \R packages interfacing with Stan, a
  copy of which can be found in the package vignettes. See
  \code{browseVignettes("rstantools")} or \code{vignette(package =
  "rstantools")}. The document is also available online at the
  \pkg{rstantools} page on the
  \href{https://cran.r-project.org/web/packages/rstantools/vignettes/developer-guidelines.html}{CRAN
  website}.
}
}
