% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reloo.R
\name{reloo}
\alias{reloo}
\title{Generic function for recomputing 'loo' for a subset of observations}
\usage{
reloo(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\description{
This is a generic function for which methods can be developed for different
fitted model objects. The purpose is to compute exact cross-validation for
problematic observations for which approximate leave-one-out cross-validation
may return incorrect results. See the methods in the \pkg{rstanarm} and
\pkg{brms} packages for examples.
}
