% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_config.R
\name{rstan_config}
\alias{rstan_config}
\title{Configure system files for compiling Stan source code}
\usage{
rstan_config(pkgdir = ".")
}
\arguments{
\item{pkgdir}{Path to package root folder.}
}
\description{
Creates or update package-specific system files to compile \code{.stan} files in \code{src/stan_files}.
}
\details{
The Stan source files for the package should be stored in:
\itemize{
  \item \code{inst/stan/src} for \code{.stan} files containing instructions to build a \code{stanmodel} object.
  \item \code{inst/stan/include} for files to be included via the \code{#include "mylib.stan"} directive.
  \item \code{inst/stan/include} for a \code{license.stan} file.
  \item \code{inst/stan/include} for the \code{meta_header.hpp} file, to be used for directly interacting with the Stan C++ libraries.
}
}
