% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_package_utils.R
\name{rstantools_load_code}
\alias{rstantools_load_code}
\title{Helper function for loading code in roxygenise}
\usage{
rstantools_load_code(path, trace = TRUE, ...)
}
\arguments{
\item{path}{Path to directory containing code to load}

\item{trace}{Whether to print file names as they are loaded}

\item{...}{Additional arguments passed to \code{\link{source}}}
}
\value{
\code{NULL}
}
\description{
Adapted from the \code{sourceDir} function defined
by \code{example(source)}.
}
