% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-functions.R
\name{LOO-prediction}
\alias{LOO-prediction}
\alias{loo_predict}
\alias{loo_linpred}
\alias{loo_predictive_interval}
\title{Generic functions for LOO predictions}
\usage{
loo_predict(object, ...)

loo_linpred(object, ...)

loo_predictive_interval(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\value{
\code{loo_predict} and \code{loo_linpred} methods should return a vector
and \code{loo_predictive_interval} methods should return a two-column
matrix similar to \code{\link{predictive_interval}}.
}
\description{
See \code{loo_predict.stanreg} in the \pkg{rstanarm} package for an
example.
}
\examples{
# Default method takes a numeric matrix (of draws from posterior
# predictive distribution)
ytilde <- matrix(rnorm(100 * 5, sd = 2), 100, 5) # fake draws
predictive_interval(ytilde, prob = 0.8)

# Also see help("predictive_interval", package = "rstanarm")

}
\seealso{
The \pkg{rstanarm} package for example methods
  (\url{https://CRAN.R-project.org/package=rstanarm}).

The guidelines for developers of R packages interfacing with Stan, a
  copy of which can be found in the package vignettes. See
  \code{browseVignettes("rstantools")} or \code{vignette(package =
  "rstantools")}.
}
