% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary}
\alias{prior_summary}
\alias{prior_summary.default}
\title{Generic function for extracting information about prior distributions}
\usage{
prior_summary(object, ...)

\method{prior_summary}{default}(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods.}
}
\value{
\code{prior_summary} methods should return an object containing
  information about the prior distribution(s) used for the given model.
  The structure of this object will depend on the method.

  The default method just returns \code{object$prior.info}, which is
  \code{NULL} if there is no \code{'prior.info'} element.
}
\description{
See \code{prior_summary.stanreg} in the \pkg{rstanarm} package for an
example.
}
\examples{
# See help("prior_summary", package = "rstanarm")

}
\seealso{
The \pkg{rstanarm} package for example methods
  (\url{https://CRAN.R-project.org/package=rstanarm}).

The guidelines for developers of R packages interfacing with Stan, a
  copy of which can be found in the package vignettes. See
  \code{browseVignettes("rstantools")} or \code{vignette(package =
  "rstantools")}.
}

