% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_summary.R
\name{prior_summary}
\alias{prior_summary}
\alias{prior_summary.default}
\title{Generic function for extracting information about prior distributions}
\usage{
prior_summary(object, ...)

\method{prior_summary}{default}(object, ...)
}
\arguments{
\item{object}{The object to use.}

\item{...}{Arguments passed to methods. See the methods in the \pkg{rstanarm}
package for examples.}
}
\value{
\code{prior_summary()} methods should return an object containing
information about the prior distribution(s) used for the given model.
The structure of this object will depend on the method.

The default method just returns \code{object$prior.info}, which is
\code{NULL} if there is no \code{'prior.info'} element.
}
\description{
See \href{https://mc-stan.org/rstanarm/reference/prior_summary.stanreg.html}{prior_summary.stanreg()}
in the \pkg{rstanarm} package for an example.
}
\examples{
# See help("prior_summary", package = "rstanarm")

}
\seealso{
\itemize{
\item The \pkg{rstanarm} package (\href{https://mc-stan.org/rstanarm/}{mc-stan.org/rstanarm})
for example methods
(\href{https://CRAN.R-project.org/package=rstanarm}{CRAN},
\href{https://github.com/stan-dev/rstanarm}{GitHub}).
}

\itemize{
\item Guidelines and recommendations for developers of \R packages
interfacing with Stan and a demonstration getting a simple package working
can be found in the vignettes included with \pkg{rstantools} and at
\href{https://mc-stan.org/rstantools/articles/}{mc-stan.org/rstantools/articles}.
}
}
