% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stantools-package.r
\docType{package}
\name{stantools}
\alias{stantools}
\alias{stantools-package}
\title{Tools for Developing R Packages Interfacing with Stan}
\description{
Tools for Developing R Packages Interfacing with Stan
}

