% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stantools-package.r
\docType{package}
\name{stantools}
\alias{stantools}
\alias{stantools-package}
\title{Tools for Developing R Packages Interfacing with Stan}
\description{
The \pkg{stantools} package provides various tools for
  developers of R packages interfacing with Stan, including functions to set
  up the required package structure, S3 generic methods to unify function
  naming across Stan-based R packages, and other tools. To get started
  building a package see \code{\link{rstan_package_skeleton}}.
}
\seealso{
The guidelines for developers of R packages interfacing with Stan, a
  copy of which can be found in the package vignettes. See
  \code{browseVignettes("stantools")} or \code{vignette(package =
  "stantools")}.

The Stan users mailing list
  (\url{https://groups.google.com/forum/#!forum/stan-users}) and the issue
  tracker for the \pkg{stantools} package
  (\url{https://github.com/stan-dev/stantools/issues}) for assistance if you
  have trouble setting up your package.
}

