% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rstan.R
\name{use_rstan}
\alias{use_rstan}
\title{Add Stan infrastructure to an existing package.}
\usage{
use_rstan(pkgdir = ".", license = TRUE)
}
\arguments{
\item{pkgdir}{Path to package root folder.}

\item{license}{Logical or character; whether or not to paste the contents of a \code{license.stan} file at the top of all Stan code, or path to such a file.  If \code{TRUE} adds the default \code{GPL (>= 3)} license provided by \pkg{rstantools}.}
}
\description{
Add Stan infrastructure to an existing package.
}
\details{
Prepares a package to compile and use Stan code by performing the following steps:
\enumerate{
  \item Update \code{DESCRIPTION} file to contain all needed dependencies.
  \item If \code{src} folder does not exist, create system file \code{src/init.cpp} to ensure Stan C++ modules are recognized by \R.
  \item Create system file \code{src/stan_files/chunks/license.stan} containing default license information.
  \item Create system file \code{inst/include/meta_header.hpp} for C++ headers used by Stan code.
}
}
